refactor(vault): split My Positions subcomponents
